+++ src/md5.c
@@ -1,13 +1,13 @@
 #include <string.h>
 #include <stdint.h>
 #include <stdlib.h>
-#include <stdio.h> // DEBUG
-#include <unistd.h> // DEBUG
+// #include <stdio.h> // DEBUG
+// #include <unistd.h> // DEBUG
 
 #include "md5.h"
 #include "ft_ssl.h"
 
-void md5_padding(uint8_t chunk[CHUNK_SIZE_TOTAL], size_t * chunk_size, long message_size)
+void md5_padding(uint8_t chunk[CHUNK_SIZE_TOTAL], size_t * chunk_size, size_t message_size)
 {
     // DEBUG
     // fprintf(stderr, "GO pad\n");
@@ -34,7 +34,7 @@ void md5_padding(uint8_t chunk[CHUNK_SIZE_TOTAL], size_t * chunk_size, long mess
     // write(2, "\n", 1);
 
     // append the original size in bits
-    uint64_t bit_size = (uint64_t)message_size * 8;
+    size_t bit_size = message_size * 8;
     memcpy(chunk + *chunk_size + 1 + zeros, &bit_size, 8);
 
     *chunk_size = 64 * (block_index + 1);
@@ -52,7 +52,7 @@ void md5_final(uint32_t *h)
     }
 }
 
-void md5_update(uint8_t * chunk, uint64_t size, uint32_t * hash) {
+void md5_update(uint8_t * chunk, size_t size, uint32_t * hash) {
 
     // initialize the hash values
     uint32_t a0 = hash[0];
@@ -61,7 +61,7 @@ void md5_update(uint8_t * chunk, uint64_t size, uint32_t * hash) {
     uint32_t d0 = hash[3];
 
     // process the message in successive 512-bit chunks
-    for (uint64_t i = 0; i < size; i += 64)
+    for (size_t i = 0; i < size; i += 64)
     {
         // break chunk into sixteen 32-bit words
         uint32_t * w = (uint32_t*)(chunk + i);
